windows pachctl build working
