+++ src/server/pfs/cmds/mount_windows.go
@@ -1,0 +1,11 @@
+// +build windows
+
+package cmds
+
+import (
+	"github.com/spf13/cobra"
+)
+
+func mountCmds() []*cobra.Command {
+	return []*cobra.Command{}
+}
